Class Clim.cli

The Object-Oriented CLI definition

method arg : a. ('a, final) arg -> ('a -> unit) -> unit

arg a k adds the argument specification a to the CLI and calls k on its final value.

method set : a. ('a, final) arg -> 'a Stdlib.ref -> unit

set a r adds the argument specification a to the CLI and sets the reference with its final value.

method virtual entrypoint : unit -> 'r

Entrypoint definition.

method man_xrefs : Cmdliner.Manpage.xref list

Inherted Cmdliner definitions.

method man : Cmdliner.Manpage.block list

method envs : Cmdliner.Term.env_info list

method doc : string

method version : string option

method name : string

method term : 'r Cmdliner.Term.t * Cmdliner.Term.info

method run : 'r

Runs the entrypoint according to the CLI definition.